+++ src/components/notes/table/NoteTable.js
@@ -66,6 +66,8 @@ function NoteTable({ apis }) {
                 <ResizableAndMovableColumn
                     dataKey={field.id}
                     label={(<strong>{field.title}</strong>)}
+                    sortBy={settingsApi.settings.noteColumnSorter ? settingsApi.settings.noteColumnSorter.field : null}
+                    sortDirection={settingsApi.settings.noteColumnSorter ? settingsApi.settings.noteColumnSorter.direction : null}
                     onResize={async data => {
                         await onResize(data, field.id, getColumnWidth(columnIndex) + data.deltaX);
 
@@ -162,6 +164,7 @@ function NoteTable({ apis }) {
                                 rowCount={dataSource.length + 1}
                                 fixedRowCount={1}
                                 cellRenderer={({ columnIndex, rowIndex, key, style }) => {
+                                    const field = sortedAndFilteredFields[columnIndex];
                                     const note = dataSource[rowIndex - 1];
                                     const classNames = [];
 
@@ -189,7 +192,7 @@ function NoteTable({ apis }) {
                                         style.backgroundColor = backgroundColor;
                                     }
 
-                                    const onClick = (event, rightClick) => {
+                                    const onClick = async (event, rightClick) => {
                                         if (note) {
                                             multiSelectionHandler(
                                                 rowData => rowData.id,
@@ -197,6 +200,42 @@ function NoteTable({ apis }) {
                                                 noteApi.selectedNoteIds,
                                                 noteApi.setSelectedNoteIds,
                                                 rightClick)({ event, rowData: note });
+                                        } else if (!rightClick) {
+                                            let direction = 'ascending';
+
+                                            if (settingsApi.settings.noteColumnSorter &&
+                                                settingsApi.settings.noteColumnSorter.field === field.id) {
+                                                if (settingsApi.settings.noteColumnSorter.direction === 'ascending') {
+                                                    direction = 'descending';
+                                                } else {
+                                                    direction = null;
+                                                }
+                                            }
+
+                                            if (direction) {
+                                                await settingsApi.updateSettings({
+                                                    noteColumnSorter: {
+                                                        field: field.id,
+                                                        direction
+                                                    }
+                                                });
+                                            } else {
+                                                await settingsApi.updateSettings({
+                                                    noteColumnSorter: null
+                                                });
+                                            }
+                                        }
+                                    };
+
+                                    const onDoubleClick = async () => {
+                                        if (note) {
+                                            noteApi.setSelectedNoteIds(note.id);
+                                        } else {
+                                            await onResize({ stop: true }, field.id, getWidthForType(field.type));
+
+                                            if (gridRef.current) {
+                                                gridRef.current.recomputeGridSize();
+                                            }
                                         }
                                     };
 
@@ -206,7 +245,7 @@ function NoteTable({ apis }) {
                                             style={style}
                                             className={classNames.join(' ')}
                                             onClick={event => onClick(event, false)}
-                                            onDoubleClick={() => noteApi.setSelectedNoteIds(note.id)}
+                                            onDoubleClick={onDoubleClick}
                                             onContextMenu={event => onClick(event, true)}>
                                             {getCellRenderer({ columnIndex, rowIndex })}
                                         </div>